requirements.txt
Django
djangorestframework
requests
black
psycopg2
coverage
